+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Note that the displayed date is in the format `dd-mm-yyyy`
 
 ### Changes
 
-- From this version, if user does not have a required permission, the bot says that instead of saying `Send w/botperms to get a list of permissions you habe given...`
+- From this version, if user does not have a required permission, the bot says that instead of saying `Send w/botperms to get a list of permissions you have given...`
 
 ## [v1.9.3]